requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
click==8.0.4
colorama==0.4.4
Flask==2.0.3
importlib-metadata==4.11.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
python-decouple==3.6
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
Werkzeug==2.0.3
zipp==3.7.0
